+++ crates/typst-syntax/src/lexer.rs
@@ -480,7 +480,7 @@ impl Lexer<'_> {
         self.s.eat_while(char::is_ascii_digit);
 
         let read = self.s.from(start);
-        if self.s.eat_if('.') && self.space_or_end() && read.parse::<usize>().is_ok() {
+        if self.s.eat_if('.') && self.space_or_end() && read.parse::<u64>().is_ok() {
             return SyntaxKind::EnumMarker;
         }
